Implemented Sending/Receiving of messages.
MainActivity:
- when active:
  - sends current user location
  - receive other users location (A)
- when in-active (in paused state, i.e., Chat is open)
  - receive other users location (A)

- (A) Receving other users location:
  - if `msgts` is set (meaning new messages where received)
    - it pull those messages and prepares thems in `SmasApp`
  - essentially ChatActivity does nothing regarding msg pull
  - when sending a message, `SmasApp` triggers `MainActivity` to pull
    again
    - so we don't have to wait the next location pull for fetching..

+++ clients/android-new/smas/src/main/java/cy/ac/ucy/cs/anyplace/smas/data/db/SmasDAO.kt
@@ -15,8 +15,7 @@ interface SmasDAO {
   @Query("DELETE FROM ${CHAT.DB_SMAS_MSGS}")
   fun dropMsgs()
 
-  // TODO getLastMessageTimestamp
   @Query("SELECT time FROM ${CHAT.DB_SMAS_MSGS} ORDER BY time DESC LIMIT 1")
-  fun lastMsgTimestamp(): Int?
+  fun lastMsgTimestamp(): Long?
 
 }
